# scripts/openapi/openapi_v0.9.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenScan3 API v0.9",
    "version": "0.9"
  },
  "paths": {
...
    "/system/repair/openscan3": {
      "post": {
        "tags": [
          "system repair"
        ],
        "summary": "Repair Openscan3",
        "operationId": "repair_openscan3",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/projects/{project_name}/scan": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Scan With Description",
        "description": "Add a new scan to a project and return the created Task\n\nArgs:\n    project_name: The name of the project to add the scan to\n    camera_name: The name of the camera to use for the scan\n    scan_settings: The settings for the scan\n    scan_description: Optional description for the scan\n\nReturns:\n    Task: The Task representing the started scan",
        "operationId": "add_scan_with_description",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Name"
            }
          },
          {
            "name": "camera_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Camera Name"
            }
          },
          {
            "name": "scan_description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "",
              "title": "Scan Description"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanSetting"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_name}/scans/{scan_index}": {
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Scan",
        "description": "Delete a scan from a project\n\nArgs:\n    project_name: The name of the project\n    scan_index: The index of the scan to delete\n\nReturns:\n    DeleteResponse: Result of the deletion operation",
        "operationId": "delete_scan",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Name"
            }
          },
          {
            "name": "scan_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Scan",
        "description": "Get Scan by project and index\n\nArgs:\n    project_name: The name of the project\n    scan_index: The index of the scan\n\nReturns:\n    Scan: The scan object",
        "operationId": "get_scan",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Name"
            }
          },
          {
            "name": "scan_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scan"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/": {
      "get": {
        "tags": [
          "openscan"
        ],
        "summary": "Get Software Info",
        "description": "Get information about the scanner software",
        "operationId": "get_software_info",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoftwareInfoResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/logs/tail": {
      "get": {
        "tags": [
          "openscan"
        ],
        "summary": "Tail Logs",
        "description": "Show or follow current logs.\n\nWhen follow=false (default), returns the last N lines of the selected log.\nWhen follow=true (text mode only!), streams new lines as they are written (like `tail -f`).\n\nArgs:\n    format: \"text\" for openscan_firmware.log, \"json\" for openscan_detailed_log.json.\n    lines: Number of last lines to return initially.\n    follow: If true, stream appended log lines in text mode.\n    poll_interval: Poll interval (seconds) when following in text mode.\n\nReturns:\n    A response with the requested log content.",
        "operationId": "tail_logs",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "text",
              "title": "Format"
            }
          },
          {
            "name": "lines",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Lines"
            }
          },
          {
            "name": "follow",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Follow"
            }
          },
          {
            "name": "poll_interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1,
              "title": "Poll Interval"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logs/archive": {
      "get": {
        "tags": [
          "openscan"
        ],
        "summary": "Download Logs Archive",
        "description": "Create and download a ZIP archive containing all log files.\n\nThe archive includes rotated files for both text and JSON logs, using\ndeflate compression for reasonable size to share e.g. via email.\n\nReturns:\n    FileResponse serving the generated ZIP. The temp file is deleted after send.",
        "operationId": "download_logs_archive",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/device/shutdown": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Shutdown",
        "description": "Shutdown system and optionally save config.\n\nArgs:\n    save_config: Whether to save the current configuration before shutting down",
        "operationId": "shutdown",
        "parameters": [
          {
            "name": "save_config",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Save Config"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Shutdown Device Shutdown Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "deleted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message",
          "deleted"
        ],
        "title": "DeleteResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PathMethod": {
        "type": "string",
        "enum": [
          "fibonacci"
        ],
        "title": "PathMethod"
      },
...
      "Scan": {
        "properties": {
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "description": "The name of the project this scan belongs to."
          },
          "index": {
            "type": "integer",
            "title": "Index",
            "description": "The sequential index of the scan within the project."
          },
          "created": {
            "type": "string",
            "format": "date-time",
            "title": "Created",
            "description": "The timestamp when the scan was created."
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "The final, persistent status of the scan, mirroring the associated Task status.",
            "default": "pending"
          },
          "settings": {
            "$ref": "#/components/schemas/ScanSetting",
            "description": "The settings used for this scan."
          },
          "camera_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Name"
          },
          "camera_settings": {
            "$ref": "#/components/schemas/CameraSettings",
            "description": "The camera settings used for this scan."
          },
          "current_step": {
            "type": "integer",
            "title": "Current Step",
            "default": 0
          },
          "system_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Message"
          },
          "last_updated": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration": {
            "type": "number",
            "title": "Duration",
            "default": 0.0
          },
          "total_size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Size Bytes",
            "description": "Total size of all files belonging to the scan, in bytes.",
            "default": 0
          },
          "stacked_size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stacked Size Bytes",
            "description": "Total size of focus-stacked JPEG files in scanXX/stacked, in bytes.",
            "default": 0
          },
          "photos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photos",
            "description": "Relative photo paths of all photos for this scan (e.g. scan01_001.jpg or stacked/stacked_scan01_001.jpg)."
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "stacking_task_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StackingTaskStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "project_name",
          "index",
          "settings",
          "camera_settings"
        ],
        "title": "Scan",
        "description": "Represents a single scan session within a project."
      },
      "ScanSetting": {
        "properties": {
          "path_method": {
            "$ref": "#/components/schemas/PathMethod",
            "description": "Scanning path generator (e.g. fibonacci or spriral).",
            "default": "fibonacci"
          },
          "points": {
            "type": "integer",
            "maximum": 999.0,
            "minimum": 1.0,
            "title": "Points",
            "description": "Number of points in scanning path.",
            "default": 130
          },
          "image_format": {
            "type": "string",
            "enum": [
              "jpeg",
              "raw",
              "dng",
              "rgb_array",
              "yuv_array"
            ],
            "title": "Image Format",
            "description": "Output image format (JPEG, DNG, RGB array or YUV array).",
            "default": "jpeg"
          },
          "min_theta": {
            "type": "number",
            "maximum": 180.0,
            "minimum": 0.0,
            "title": "Min Theta",
            "description": "Minimum theta angle in degrees for constrained paths.",
            "default": 12.0
          },
          "max_theta": {
            "type": "number",
            "maximum": 180.0,
            "minimum": 0.0,
            "title": "Max Theta",
            "description": "Maximum theta angle in degrees for constrained paths.",
            "default": 125.0
          },
          "min_phi": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Phi",
            "description": "Optional minimum phi angle in degrees for constrained paths.",
            "default": 0
          },
          "max_phi": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Phi",
            "description": "Optional maximum phi angle in degrees for constrained paths.",
            "default": 360.0
          },
          "optimize_path": {
            "type": "boolean",
            "title": "Optimize Path",
            "description": "Enable path optimization for faster scanning.",
            "default": true
          },
          "optimization_algorithm": {
            "type": "string",
            "title": "Optimization Algorithm",
            "description": "Path optimization algorithm to use.",
            "default": "nearest_neighbor"
          },
          "focus_stacks": {
            "type": "integer",
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Focus Stacks",
            "description": "Number of photos with different focus per position.This ignores AF and you need to set a focus range.Focus values will then be evenly spaced between min and max.",
            "default": 1
          },
          "pause_before_capture_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pause Before Capture Ms",
            "description": "Pause in milliseconds before capture to let vibrations settle.",
            "default": 0
          },
          "focus_range": {
            "prefixItems": [
              {
                "type": "number",
                "maximum": 15.0,
                "minimum": 0.0
              },
              {
                "type": "number",
                "maximum": 15.0,
                "minimum": 0.0
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Focus Range",
            "description": "Minimum and maximum focus distance in diopters.",
            "default": [
              10.0,
              15.0
            ]
          }
        },
        "type": "object",
        "title": "ScanSetting"
      },
...
      "SoftwareInfoResponse": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Scanner model identifier, if configured."
          },
          "firmware_version": {
            "type": "string",
            "title": "Firmware Version",
            "description": "Currently running firmware version string."
          },
          "last_shutdown_was_unclean": {
            "type": "boolean",
            "title": "Last Shutdown Was Unclean",
            "description": "Indicates whether the previous shutdown finished cleanly."
          },
          "runtime_dir": {
            "type": "string",
            "title": "Runtime Dir",
            "description": "Absolute path used for runtime state files."
          },
          "runtime_disk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiskUsage"
              },
              {
                "type": "null"
              }
            ],
            "description": "Disk usage snapshot for the runtime directory filesystem."
          },
          "projects_disk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiskUsage"
              },
              {
                "type": "null"
              }
            ],
            "description": "Disk usage snapshot for the projects directory filesystem."
          },
          "uptime_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uptime Seconds",
            "description": "Current system uptime in seconds, if available."
          }
        },
        "type": "object",
        "required": [
          "firmware_version",
          "last_shutdown_was_unclean",
          "runtime_dir"
        ],
        "title": "SoftwareInfoResponse",
        "description": "Information block served by /next/openscan."
      },
...
      "Task": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "is_exclusive": {
            "type": "boolean",
            "title": "Is Exclusive",
            "description": "Whether this task is exclusive and should not run concurrently",
            "default": false
          },
          "is_blocking": {
            "type": "boolean",
            "title": "Is Blocking",
            "description": "Whether this task is blocking and should run in a separate thread",
            "default": false
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "pending"
          },
          "progress": {
            "$ref": "#/components/schemas/TaskProgress"
          },
          "result": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "run_args": {
            "items": {},
            "type": "array",
            "title": "Run Args",
            "description": "Positional arguments the task was started with."
          },
          "run_kwargs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Run Kwargs",
            "description": "Keyword arguments the task was started with."
          }
        },
        "type": "object",
        "required": [
          "name",
          "task_type"
        ],
        "title": "Task",
        "description": "Represents a background task."
      },
...
    }
  }
}